Fix DS-Operator version
This commit adds changes updating ds-operator version as the binary folder provided by asigra was named 14_2_0_3 whereas the ds-operator version in reality is 14_2_0_2. After syncing with Asigra, they confirmed that the latter is the correct version and this change aims to address this concern so that pkg correctly reports the correct ds-operator version.